renamed main nodelet to node
